Move `re_video::Time` to own file
